test(workflow): movido test de IT a behavior
