+++ src/test/java/es/in2/desmos/it/application/PublishWorkflowBehaviorTest.java
@@ -9,6 +9,8 @@ import es.in2.desmos.domain.models.BrokerNotification;
 import es.in2.desmos.domain.repositories.AuditRecordRepository;
 import es.in2.desmos.domain.services.api.QueueService;
 import es.in2.desmos.infrastructure.controllers.NotificationController;
+import es.in2.desmos.objectmothers.AuditRecordMother;
+import es.in2.desmos.objectmothers.BrokerNotificationMother;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,8 +19,11 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import reactor.test.StepVerifier;
 
 import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @Testcontainers
@@ -176,6 +181,28 @@ class PublishWorkflowBehaviorTest {
         } catch (Exception e) {
             log.error("Error: " + e.getMessage());
         }
+    }
+
+    @Test
+    void itShouldCreateAuditRecordWhenBrokerNotificationDataIsAnArrayOfEntities() {
+
+        var inputBrokerNotification = BrokerNotificationMother.withDataArray();
+        var expectedCreatedAuditRecordsList = List.of(
+                AuditRecordMother.createAuditRecordFromDataMap("0", inputBrokerNotification.data().get(0)),
+                AuditRecordMother.createAuditRecordFromDataMap("0", inputBrokerNotification.data().get(1))
+        );
+
+        notificationController.postBrokerNotification(inputBrokerNotification).block();
+        pendingPublishEventsQueue.getEventStream().subscribe();
+        var auditRecordsListMonoResult = auditRecordRepository.findAll().collectList();
+
+        StepVerifier
+                .create(auditRecordsListMonoResult)
+                .consumeNextWith(result ->
+                        assertThat(result)
+                                .usingRecursiveFieldByFieldElementComparatorIgnoringFields("id", "processId", "createdAt", "hash", "hashLink")
+                                .containsAll(expectedCreatedAuditRecordsList))
+                .verifyComplete();
     }
 
 }
